Created a base namespace and moved GridDistributions there
